Update 2.0_Getting_Started.md
Condensed some parts down to go with the less is more philosophy.

In No.3 above, was a little confused here as it makes python sound inferior to R. Does R have more PA examples than python? and did you mention this simply to point out that you wish to expand PA examples in python in this crazy world we live in? 

Also this goes more so for the intro section but just thought of it..do you think this book would also benefit those who are new to or are interested in switching to the HR analytics space? could potentially add that